feat: etc API 정리
