+++ src/api/etc/index.ts
@@ -1,0 +1,2 @@
+export { default as useRegionInfo } from "./useRegionInfo";
+export * from "./etc";
